day 6 part 2

+++ src/main/java/Main.java
@@ -9,7 +9,7 @@ public class Main {
     public static void main(String[] args) {
         String file = "C:\\Users\\Lenovo\\Desktop\\study\\AdventOfCode2023\\src\\main\\resources\\day6\\input.txt";
         try {
-            int result = BoatGame.numberOfWaysToWin(file);
+            long result = BoatGame.numberOfWaysToWin(file);
             System.out.println(result);
         } catch (IOException e) {
